first attempt at supporting button hold
previously the button handlers would get somewhat
confused when pressing and holding or releasing multiple
buttons at once (e.g. an arrow key and space bar)

making it work more like an arcade joystick so holding
the stick (arrow key) means keep accelerating.

don't love this implementation so far but it is ok for now
would like to get the style to be more parallel to how the other
character update systems work